Add checks for Debian (/etc/debian_version)

+++ create_ssl.sh
@@ -22,11 +22,14 @@ fi
 
 # Skip install Certbot, if needed
 if [[ $2 != "--skip-install" ]] || [[ $3 != "--skip-install" ]]; then
-    # Install apt-add-repository package
-    sudo apt install software-properties-common -y
+    # If OS is not Debian
+    if [[ ! -f /etc/debian_version ]]; then
+        # Install apt-add-repository package
+        sudo apt install software-properties-common -y
 
-    # Add repository for Ubuntu universe, Certbot
-    sudo add-apt-repository -y universe && sudo apt-add-repository -y ppa:certbot/certbot
+        # Add repository for Ubuntu universe, Certbot
+        sudo add-apt-repository -y universe && sudo apt-add-repository -y ppa:certbot/certbot
+    fi
 
     # Update
     sudo apt update
